Refuse to format-at-offset when the offset does not point to the last character in a token
Summary:
Since our Nuclide integration does not reposition the cursor after as-you-type formatting, and hackfmt does not print partial tokens, the behavior of as-you-type formatting is undesirable when the cursor is within a token (such as a string literal). The formatting works correctly (minus changes after the end of the token at the given offset), but the cursor is repositioned to the end of the inserted text (and thus jumps to the end of the string literal).

This change just has hackfmt throw an exception when an offset is requested which is not at the end of a token.

Reviewed By: arxanas

Differential Revision: D7553254

fbshipit-source-id: 69e3d109a2d670854b11ddb1d75d97e30c8e6daa