Remove classname from is-expressions
Summary:
Because classnames are opaque type aliases, they don't exist at runtime. This means supporting them in is-expressions would be unreliable, as any custom desugaring we did would only apply if the classname hint were used directly; not hidden behind a type alias.

Therefore, we'll have to punt on supporting classname in `is`-expressions until we choose to lower it into the runtime as a real type.

This is probably for the best, as the team is investigating how to represent first-class classes in a more flexible way than strings.

Reviewed By: oulgen

Differential Revision: D7561469

fbshipit-source-id: b39bb48138f36d0346157bfbc54a5117b3e4422b