Have front-ends set immutable-related flags on classes and props
Summary:
Use the __Const attribute on classes and properties to set the is_immutable, has_immutable, and no_dynamic_props attributes on classes, and the is_immutable attribute on properties.
Add unit tests for all the basic write paths.
Support for constructors, verification, and additional unit tests are coming in later diffs.

Reviewed By: oulgen

Differential Revision: D7567596

fbshipit-source-id: 93ca47dcbc2563c61d6696e3dffd3b6e60913140