Avoid unnecessary Tunresolved's when integrating local environments
Summary: When integrating local environments, we wrap the child types of a local variable into `Tunresolved` and unify them.  This wrapping into `Tunresolved` is unnecessary when there is only one child type because there is nothing to unify with.  However, wrapping the type into an unresolved allows it to grow in the cases where it shouldn't.  See a new test case below.

Differential Revision: D7569454

fbshipit-source-id: 3609c2903868f07c7a958b7f4d88cc0537dc943f