Allow function subtypes to return mutably owned values if supertype returns immutable values
Summary: as in title, this should be safe since mutably owned values has only one references and when such value is returned - this single reference is passed to caller. On the caller side this reference can be safely treated as immutable because we know that there are no other mutable aliases to the same value.

Reviewed By: KendallHopkins

Differential Revision: D7561133

fbshipit-source-id: 6055ef8ca5c7511d767d9cc6d4da5570c9dde0c3