Type array_filter in TAST
Summary: Currently, the identifier `array_filter` is not annotated with a type in the TAST, since its type is rewritten depending on its arguments. This change annotates it with its function type (retreived from hhi), but with the return type replaced with the specific return type used at that call site.

Reviewed By: jamesjwu

Differential Revision: D7495885

fbshipit-source-id: b5cafe2fcd525802c1417845d192628b31449cae