Type AFvalues occurring in key position in TAST
Summary: Currently, we leave AFvalues occuring in key position without a type annotation. Since this shouldn't happen, the type annotation  we choose is not very important. In the interest of making every expression explicitly typed, this change annotates the AFvalue node with the type we currently give the expression in the typechecker (`Tprim Tint`).

Reviewed By: andrewjkennedy

Differential Revision: D7495735

fbshipit-source-id: 3a4ff38074781193f0abbd5a041090978dd106cb