Annotate Obj_get IDs with the type of the entire Obj_get expression in TAST
Summary:
Currently, the `Id` in a non-dynamic `Obj_get` is not annotated with a type in the TAST. This is because it doesn't really have a meaningful type. For example, consider the ID `foo` in the function `test`:

```
class C {
  public ?int $foo;
}

function test(C $c): ?int {
  return $c->foo;
}
```

The type of `$c->foo` is `?int`, but the identifier `foo` itself arguably has no meaningful type. We are in the position of needing to assign a type to it because dynamic `Obj_get` expressions (which may have an expression rather than an identifier, e.g. `$c->$foo`) are not distinguished from static ones in the AST.

In the interest of ensuring that every expression has a type, we would like to eliminate this implicitly typed expression. The least wrong type we can assign the identifier is probably the type of the accessed property or method, and this diff makes that change.

Reviewed By: pittsw

Differential Revision: D7495798

fbshipit-source-id: 1651c644901dd786895cfc45de6aa304c2c3ffcd