Type special functions as Tany in TAST
Summary:
In the interest of making every expression explicitly typed, this change types special functions as `Tany` rather than leaving them without a type.

Ultimately, we should annotate these special functions with the appropriate function type specific to each call site, but this can be done in a future diff.

Reviewed By: ljw1004

Differential Revision: D7495897

fbshipit-source-id: de17c220e3edef2a602e6ca7c4dd0f560152b46b