Type array_map in TAST
Summary:
Currently, the identifier `array_map` is annotated with Tany in the TAST, since its type is rewritten depending on its arguments. Although we retrieve the function type from its definition in hhi and rewrite it to the appropriate type for that call site, we did not make use of it when constructing the TAST node.

This change simply carries that function type into the node's annotation rather than leaving it annotated as Tany.

Reviewed By: manzyuk

Differential Revision: D7593407

fbshipit-source-id: 99a12d14d2cbacd9b748dec17485d238db66591f