reject copy() for directories
Summary:
Check that the file being copied is a regular file before trying to copy it.

This ensures that if a directory or other special file is being copied that we
fail before creating the destination path.  Previously copy() would
successfully open a file descriptor for the source directory, would create the
destination as an empty file, and only fail when it tried to read() from the
directory file handle.

This fixes part of the bad behavior reported in
https://github.com/facebook/hhvm/issues/8175

Reviewed By: markw65

Differential Revision: D7592426

fbshipit-source-id: b1902b212e06ec55b16091ce59e4309a46eff8b5