Change Tast.ClassIdAnnotation.t from ty option to ty
Summary: Now that ClassIdAnnotations are constructed in only two places (when we construct `parent::__construct()` nodes and in `static_class_id`), it is easy to see that we no longer construct any class ID nodes with an annotation of `None`. This change makes the type annotation on class IDs non-optional.

Reviewed By: pittsw

Differential Revision: D7495489

fbshipit-source-id: e7b0ed8e84fa9ae6683de12995b557b7a20faddb