Model unwind edges more precisely in HHBBC
Summary:
HHBBC currently uses a "factored" exit representation, which is simple but
conservative. That is, it adds edges which cannot be actually traversed at
runtime. This is fine since it only means that types may be unioned together
where they wouldn't have to. However, I wish to track iterator liveness exactly,
and this causes problems. IE, this conservative representation may claim a throw
can jump past an IterFree instruction (when in fact it cannot).

Resolve this issues by modeling exceptional edges more precisely. Instead of
having a single set of factored edges for a block, split it into throw edges and
unwind edges. Throw edges are similar to the current factored edges. They
represent edges which can be traversed because of a throw, and generally can
happen anywhere within the block. Unwind edges represent edges that can be
traversed via unwinding from an Unwind instruction. In general, these edges can
be different from throw edges and thus are modeled separately.

The calculation of unwind and throw edges are done when the function is
parsed. Since modeling the control flow of the unwinder is tricky, we do it via
simulation of the unwinder. The unwinder is a state machine, so we start with
its known initial states. For each iteration, we compute the new states
resulting from the known states by applying the unwinder rules, and add them to
the known states. We repeat this process until we reached a fixed-point (no new
states). The total closure of all the possible unwinder states gives the
possible edges between regions (and thus the associated blocks).

Reviewed By: markw65

Differential Revision: D7554600

fbshipit-source-id: 27f4fade93a783650149785d0871be1993e8c3e7