Propagate known refcounts through phi nodes
Summary:
We know the lb of the output of a DefLabel is the min of the lower
bounds of its inputs, so bump it up using unsupported_refs.

This doesn't allow us to fold inc/dec pairs across the phi node, but
does get us a bit closer, and more information is always better.

Reviewed By: ottoni

Differential Revision: D7601541

fbshipit-source-id: 52d82b31e7186dc78c41d904d6a9420eddb00e95