initial import of cppo
Summary:
initial import of cppo

potential issues:

I named it cppo for simplicity, not hh_cppo or similar.

This is because you can invoke ppx's from the command line
with just an executable, so I don't think modularity is all that
necessary.

This could become a potential issue later, I'm open to suggestions.

I had to make some minor adjustments to get it to build.

The original lines in 1.6.4 have been preserved in comments. It should
be straightforward to follow.

The version module is generated. I just named mine "1.6.4 with changes"

a little weird but whatever.

Reviewed By: dabek

Differential Revision: D7464465

fbshipit-source-id: b13a16be680ac93a3598a798a39e9efd1e80a7cc