Fix reflection output for d/varray parameters
Summary:
The output of getDefaultValueText() for parameters that d/varray default values
was very inconsistent, both between HPHPc and HackC, and depending on the
initial value. Fix this by having the PHPOutput mode of the serializer print out
d/varrays as themselves (not as arrays) and have HackC do the same.

However, this exposed other issues. Namely, in HackArrDVArrs mode, the parser
turns d/varrays into dicts and vecs, which means getDefaultValueText() doesn't
return what the source code actually wrote. To fix this, the d/varray
distinction needs to be maintained into the emitter. We introduce the concept of
a "DVOverride" for the serializer/unserializer. This is a side table which
records whether a vec/dict is actually a varray/darray. This can then be fed
into the various parts which care about that distinction. Likewise, HackC should
serialize d/varrays as themselves even when HackArrDVArrs is enabled. This lets
the HHAS parser track these distinctions and generate the default value text
properly.

With these changes, the output of getDefaultValueText() (and other similar
reflection) should always reflect d/varrays as written, rather than what they
actually are at runtime.

One downside of this approach is that its not robust to round-tripping. That is,
if you dump the repo to HHAS, and then re-parse it, the reflection text won't
stay the same (because we've lost whether the values were d/varrays originally).

Reviewed By: alexeyt

Differential Revision: D7558603

fbshipit-source-id: 869f6655f212b0826ebd5b9a4219cc0393f47018