Fix Unsafe bug due to optimisations for Noop
Summary:
Earlier optimisations to prune work for `Noop` cases seems to have eaten a
scenario for `Unsafe`;
```
<?hh // strict

class A {
  final public function f() : C {
    //UNSAFE_BLOCK
  }
}
```
This has a `Missing` (because `SyntaxList`s are expressed as `Missing` when empty) for the `ComponentStatement` that is the body of `f`. In that case, we don't process the compound statement in the lowerer and thus miss the `//UNSAFE_BLOCK` pragma.

Reviewed By: CatherineGasnier

Differential Revision: D7635509

fbshipit-source-id: ee7f488346043a137f59135f96780911a827fadb