Using local config to gate FFP cutover
Summary:
This diff, commandeered from jamesjwu, adds facilities to gate the cutover to the
full fidelity parser. When `/etc/hh.conf` contains
```
use_full_fidelity_parser = true
```
we should use the FFP (next diff), if not, stick to `Parser_hack`. This way,
we're not tied to the release cycle if we want to revert.

Reviewed By: dabek

Differential Revision: D7518255

fbshipit-source-id: 9441bc7374704adb52259d43fd62a6f09ce3af06