Only show completely untypechecked code in typecoverage
Summary:
So I'm ambivalent about this change: on one hand, it's noisy to send partially typechecked lines to the user (especially since "null" is considered partially typed, which is very unactionable).

On the other hand, it's weird if you see under 100% on your coverage percentage, but when you click the button you don't see any feedback to fix it.

IMO, I think it's good to just turn it off for now until we can fix it by using the Typed AST for type coverage.

Feedback would be appreciated!

Reviewed By: vassilmladenov

Differential Revision: D7623107

fbshipit-source-id: b8a5a586c15eee5494cf52d56d906bb43a7c3467