Omit AnnotatedAST module name when pretty-printing NAST/TAST
Summary: By default, the ppx_deriving show plugin will include module/functor names in the output. The AnnotatedAST functor name doesn't add any value when printing the NAST/TAST, so this change configures the show plugin not to include it (it is still printed for record types, unfortunately, but this behavior does not seem to be configurable).

Reviewed By: ljw1004

Differential Revision: D7644282

fbshipit-source-id: 5524ec19b1d74476a457eef5e5a9ee1481e43253