Don't indent multiline concatenations passed as function arguments
Summary: Indenting these continuations makes it easier to distinguish these expressions from one another when they are in a list (e.g., a function call's arguments list or a vec literal), but it is conventional not to indent continuations of multiline strings in Hack. To conform to this convention, we make concatenation an exception to this indentation behavior.

Reviewed By: arxanas

Differential Revision: D7643656

fbshipit-source-id: af43dcaa662f78dddd1028c10d32b67ec41d4d2f