Don't fall over if jemalloc doesn't support background threads
Summary: On OSX jemalloc (even version 5) doesn't support background threads; the only way to tell is that mallctl with the background_thread option will fail. We already deal with this when trying to set the background_thread option by passing true for errOk, but we forgot to do this when we first read it.

Reviewed By: markw65

Differential Revision: D7647523

fbshipit-source-id: 09b5d5c14c061f3c208db8b54adcf0f529203a9a