Fix unification of Tvar's
Summary:
This diff started from my investigation of a false type checker error.  I have added the code example that triggered it as a new test case (see `lambda_apply2.php` below).  I managed to traced the issue to a bug in the unification algorithm: in some cases when we unify two type variables bound to unresolved types the result type comes out as not unresolved, which prevents it from growing.  I've added a really crisp example of this as a unit test (see `unify_test.ml` below).  Let me walk you through it.

Suppose that we are trying to unify `#1(int)` and `#2(#1(int))` (i.e., we are unifying type variable `#1` bound to an unresolved type `(int)` in the type environment with type variable `#2` bound to an unresolved type `(#1)` in the type environment).  Unification of type variables works as follows:

(1) First, we create a fresh identifier, say `#3`, and add substitutions `#1: #3` and `#2: #3` to the substitution map.  We end up with the following schematic picture:
```
 \  /
  #3
```
Note that as a result of this operation `#1` is no longer bound to `(int)`: to get the type it is bound to we need to follow the chain of substitutions and get the type for the last variable, `#3` in our case, from the type environment.  Because `#3` is not bound to any type yet, we get `_` (the any type).

(2) Next, we unify the types `(int)` and `(#1_)` bound to the variables.  These are two singleton unresolved, so the result is going to be a singleton unresolved containing the result of unifying `int` with `#1_`.  The latter is unification of a type variable with a non-variable type.  It works as follows:

(2a) Create a fresh identifier, say `#4`, and add a substitution from `#1` to `#4`.  Because `#1` is not a terminal variable, we actually have to traverse its chain of substitutions and add a substitution for the last variable, so that we end up with the following picture:
```
 \  /
  #3
   |
  #4
```

(2b) Next, we unify the types `int` and `_`, which gives us `int`, and bind the result to `#1`, which actually again involves traversing the chain of substitutions and adding a binding for the last variable, `#4` in this case.  So, we end up with:
```
 \  /
  #3
   |
  #4
   |
  int
```

(3) The result of the above unification came out as `(#4int)`, which is the type we want to bind to `#3`, the fresh variable we introduced at the beginning.  Again, this involves traversing the chain of substitutions starting at `#3` and adding a binding for the last variables, i.e., `#4`.  So, we need to add a binding from `#4` to `(#4int)`.  That's where the bug lies: we treat this case similarly to binding `#4` to `#4int`, i.e., we falsely say that `#4` occurs at the top level of the type  we want to bind to it, and hence no new binding is necessary.  End of computation.  We end up with the above picture in which `#1` and `#2` are bound to `int` rather than `(int)`.

The fix I came up with is as follows: don't treat `#4` as occurring at the top level of `(#4int)` (but still elide singleton unresolved's inside an option).  I've added a unit test case for this.  With this fix, the code in `lambda_apply2.php` type checks.

Reviewed By: andrewjkennedy

Differential Revision: D7653566

fbshipit-source-id: 9c69e15224c96c0f8646a481f170d2dcc60c5c9e