Support type consts in IsTypeStruct
Summary:
When resolving type structures, we may encounter references like `this`, `self`, and `parent`. Ordinarily, these are easy to resolve, because we have the declaring class and the called class for the type constant. But when resolving the anonymous type structures generated by `is`/`as` expressions, we don't have a LSB class and declaring class—the type structures are inline.

Because they could still contain the same references as declared type constants, we manually determine the declaring and called class by essentially calling `self::class` and `static::class` before resolving. If we're not in a class, these will be set to null and the appropriate fatal errors will be thrown.

A minor complication here is that `this` is checked differently from typehints. In typehints, we enforce that `this` means the value is exactly the LSB class, not a subtype. Here, because we're working with a resolved type structure, we don't know where the class name came from, so we'll just check that it's an instance of the class. I don't think we'll be able to fix that without some other changes to type structure.

Reviewed By: ricklavoie

Differential Revision: D7523950

fbshipit-source-id: 6f969c4c7826770740c657d9f3be912642286c0b