Remove 'this' is-expression desugaring from HackC
Summary:
Removes the remainder of the desugaring from HackC and defers to HHVM to optimize it as necessary.

This does mean that we lose the behavior of `this` being exactly the LSB class (not a subtype), but now the behavior will be consistent across the feature and the fix will be in one place if we choose to move forward with it.

Reviewed By: jamesjwu

Differential Revision: D7609016

fbshipit-source-id: eba3032f6ffdcb4a0ff5f6d6916e4b401331065a