Extend IterBreak to optionally include a local id
Summary:
"LIters" (about to be introduced) have an associated local (which
contains the iterator's base, not the iterator itself). This means any operation
on the iterator needs to specify the local, including IterBreak. For this
purpose, extend the representation of the IterBreak immediate. Instead of pairs
of IterKind and iterator id, its now IterKind, an iterator id, and, if IterKind
== LIter, a local id. Update all the encoding decoding logic to handle a
variable length tuple, which involves making the IterBreak immediate its own
immediate type.

Reviewed By: jano

Differential Revision: D7625299

fbshipit-source-id: 5b238648805e44ed3adb65b7536d416c5a131572