ServerEnv.Full_check_started
Summary:
`ServerEnv.needs_full_check` is a boolean - recheck is either done (false), or not (true).

If it's true, you can start a typecheck, and expect it to be false after you are done. When we'll start interrupting type-checks, there will be a third state that we want to recognize - it was started, but not finished.

I also utilize it here (and for now, only here) to remove `ServerMain.initial_check` - I just set `Full_check_started` and check for it in the main loop recheck loop to achieve (hopefully) the same result.

Reviewed By: gregory-nisbet

Differential Revision: D7550949

fbshipit-source-id: 4354479f7c2f12f85117df53361e71ec96e488f3