Update ServerTypeCheck environment before starting typing
Summary: We parsed all new files, updated necessary declarations, and could start processing commands **while** typechecking is going on. But most commands require the environment, which we update only after typechecking. Here I add an intermediate step that updates the parts of the environment that are ready by this point.

Reviewed By: alexchow

Differential Revision: D7550952

fbshipit-source-id: 60b98677d18faa0534f725148526a50e2bcfa993