Replace JobQueueDispatcher::addWorkers() with saturateWorkers()
Summary:
In a world where the maximum thread count can be modified, the
addWorkers() API bleeds the boundary between the JobQueueDispatcher and
the client---we have to know the client's intent in terms of growing
m_currThreadLimit in order to set it correctly when we adjust
m_maxThreadLimit.

To avoid this, just replace addWorkers() with a function that raises the
effective limit to the maximum.  In practice, this is the only use case
we had for this function anyway.

Reviewed By: markw65

Differential Revision: D7677375

fbshipit-source-id: e79325563b8aef5760ae1066a461da6b670258a0