LSP supports file edits
Summary:
Previously persistent_client stored for each client the set of files it had open, responding to didOpen/didClose events.

Now for LSP it also stores the content of those files, initialized in didOpen and updated in didChange; for legacy IDE, it only stores a dummy placeholder.

I've updated the commandHandler for textDocument/definition to base its results on the current content of the file. As I add new commandHandlers, they will likewise operate on the current content.

In a future diff, I plan to have it report syntax errors based on the current content as well, updated as-you-type; typing errors will continue to be based on the files on disk.

ALSO: I improved the handling of errors in builder.js. Now, if the LSP server responds with an error, then that error will show up in the ideMessages list (rather than failing the test). This way we can write tests that assert that the LSP server responded with an error.

ALSO: I did some refactoring because updating file content is needed in both lspCommand and in commandHandler. In the refactoring, made the 'Error' result also return callstacks.

Reviewed By: gabelevi

Differential Revision: D7337139

fbshipit-source-id: 7eaf4849b755d212b96fb81a23e55e631b6fed4d