Allow multiple stdout hooks on ExecutionContext
Summary: Currently only 0 or 1 stdout hooks can be installed. When the debugger installs a stdout hook to redirect output over to the debugger client, this conflicts with the hook installed by the CLI server to redirect output to the CLI client. This diff makes it possible to have multiple hooks installed, each of which receives stdout on each write.

Reviewed By: markw65

Differential Revision: D7579020

fbshipit-source-id: 5e40f9f38fdfbf379949cced2881386c4cb1f62b