Fix compile error in mysql extension
Summary:
The following compile error occurs when building with my_bool defined as bool:
   error: dereferencing type-punned pointer will break strict-aliasing rules [-Werror=strict-aliasing]

This is because we are trying to read a variable as a bool, and it is generally not supported. For reference, this is how the field is populated:
https://github.com/mysql/mysql-server/blob/8.0/libmysql/libmysql.cc#L2572

To fix, just compare the value with 0 to see if the bool was set or not.

Reviewed By: abal147

Differential Revision: D7678435

fbshipit-source-id: a06df837e319e64ce8c08e77735ec2eacfd8779a