add dump_apc_prefix to admin server
Summary:
Dumping all of APC can crash a box. It holds the lock for a long time and can produce gigabytes of output. When debugging, there is usually a particular class of keys one would like to inspect to figure out what went wrong.

This new command lets you specify a `prefix` and an optional `count` to dump keys matching a single prefix. This is substantially faster and less crash-y for targeted debugging on live hosts.

Reviewed By: asukhachev, markw65

Differential Revision: D7689894

fbshipit-source-id: e21f0fde702973043bbcf1e0bc36810ad4b83f76