get in memory deptable size RPC
Summary:
When writing an incremental saved state, we have to update an
existing saved state by inserting the entries from the in-memory
deptable.

The more entries we have in-memory, the more likely we will end
up have extraneous edges in the resulting saved state.

Exposing this number to clients will help them decide whether or not
they'd like to make an incremental saved state, or a fresh one.

Reviewed By: gregory-nisbet

Differential Revision: D7665205

fbshipit-source-id: c41e3df34eca85fae7cf5c9323ad8bd7e8571f59