Fix naming for `for` loops
Summary:
If a `for` loop is skipped, the iteration step is not executed. Had there been a definition of a new variable in the iteration step, it is not ensured to be assigned. An equivalent `while` loop will report errors of similar kind, which is intended behaviour. See examples below.

```
function foo(): void {
  for (; false; $x = 10) {}
  echo $x; // should error here
}
```
```
function bar(): void {
  while (false) {
    $x = 10;
  }
  echo $x; // error
}
```

In addition, the block should be named before the iteration.

```
function foo(): void {
  for (; true; $x = 10) {
    $y = $x; // should error here
  }
}
```
```
function bar(): void {
  while (true) {
    $y = $x; // error
    $x = 10;
  }
}
```

Reviewed By: CatherineGasnier

Differential Revision: D7686350

fbshipit-source-id: 27b91367361624dbbec0ecbdc98e5b0e3e61d517