Shrink Jcc sequences, and forward branches
Summary:
Addresses that were recorded as needing to be "far" in the relocator were dropped causing an infinite relocation loop if this was ever needed.

This adds shrinking of forward branches.
  - This will reduce the number of indirect branches, and the number of instructions needed to execute the branches.

This adds inverting and shrinking of Jcc sequences.  All Jcc sequences are emitted in a smashable form.  This will collapse them to a relative conditional branch.
  - This will improve the fallthrough rate on conditional branches hopefully improving the misprediction rate.
  - This will reduce the number of branches and instructions needed to execute such sequences.

Depends on D7622217

Reviewed By: ottoni

Differential Revision: D7708531

fbshipit-source-id: 026b6cba833bba99a4016a4b90d2893a3af507e7