Don't use c++ streams for jit serialization
Summary:
C++ streams can be slow; they're also overkill for what we're doing.

In addition, we want to be able to serialize the jit state while crashing, and apart from one hash-table that we build during serialization (which we can probably eliminate), this was the only thing thats not signal handler safe.

Reviewed By: alexeyt

Differential Revision: D7719797

fbshipit-source-id: 2d64c0874115c5b5a44eb27a32902acff6056e7b