HHVM Debugger: Improve user messaging when startup document fails to load on dummy request
Summary: If the dummy fails to hphp_invoke() the startup document, provide better error messaging and send it to the UX as a warning instead of an error. Error makes the user think the debugger cannot work at all, and in this case it can, but may not have loaded symbols referenced in the startup document.

Reviewed By: Goom11

Differential Revision: D7688374

fbshipit-source-id: ace3e0818960baa81e68a852e18f55e6dcf832db