Delete String_sequence, guess buffer size to avoid reallocations
Summary:
We have `String_sequence` type for efficient prepending/appending of the text but we've never actually used this aspect of it - in all cases we first constructed `String_sequence` value and then immediately print it. This diff removes `String_sequence` and replaces its usages with buffer - as a result it reduces amount of intermediate allocations during printing step.
Also this diff pre-allocates the buffer for the printing result to avoid repeated reallocations in case of large files

Reviewed By: andrewjkennedy

Differential Revision: D7671166

fbshipit-source-id: c96ca1a59fefd22624e6bfcd93c3bc9cba12f7aa