Fix the recent fix of unification of Tvar's
Summary: I fixed unification of `Tvar`s in D7653566 but made a mistake: when solving unification of `#n : t` with `(#n | t1 | t2 | ... | tk)`, we were supposed to introduce a fresh variable `#m`, bind it to the singleton unresolved `(t)`, and bind `#n` to `(#m | t1 | t2 | ... | tk)`.  Instead the code bound `#n` to `((t) | t1 | t2 | ... | tk)`.  Not sure if this //really// matters because in the former, correct case `#m` occurs only under the unresolved bound to `#n` and so probably cannot grow independently, but for consistency I've chosen to bring the code with agreement with the comments.

Differential Revision: D7686261

fbshipit-source-id: ec789c4a0cb5e4b63356d89e3669d8bcdff3b8f4