Improve unification of #n with ?#n
Summary:
I've just realized that when unifying `#n` with `?#n`, we forget the type `#n` was previously bound to and rebind it to `?#m` where `#m` is a fresh unresolved type.  I was able to come up with an example (added below as a test case) that exposes this problem:
```
<?hh // strict

function my_compact<T>(Vector<?T> $vector): Vector<T> {
  //UNSAFE
}

function test(bool $b): ?string {

  $nullable = Vector { 42, null };
  $non_nullable = my_compact($nullable);

  if ($b) {
    $res = $nullable;
  } else {
    $res = $non_nullable;
  }

  return $res[0];
}
```
This type checks.  Sticking `hh_show()` suggests that the type of `$nullable` is inferred as `Vector<^(int | ?^_)>`, the type of `$non_nullable` is inferred as `Vector<^(int)>`, whereas the type of `$res` is inferred as `^(Vector<^?^[unresolved]>)`.  That `^[unresolved]` allows the type of `$res` to grow when it is compared against the expected return type `?string`.

The fix is similar to what I previously suggested for unification of `#n` with an unresolved containing `#n`: bind `#n` to `?#m(t)` where `t` is the type originally bound to `#n`.

Reviewed By: CatherineGasnier

Differential Revision: D7707562

fbshipit-source-id: 54e72565e8ffb71d8afa80abd76fbb3bd29b2ada