Use snake_case everywhere
Summary:
Currently, we're using a mixture of snake_case, lowerCamelCase, and dash-case in different situations. It's probably best to use only a single one of those styles, so as to avoid confusing the consumers of the API.

I chose snake_case because it already has precedence, given the JSON output of the full-fidelity parser, and it's also usable for identifiers in code (unlike dash-case in Hack or Flow).

Differential Revision: D7717960

fbshipit-source-id: 4c5a5b75d1267d75980ccdafbfe6cace1c371f8c