Omit module names when pretty-printing NAST/TAST
Summary:
In {D7644282}, we omitted the AnnotatedAST functor name from AAST pretty-printing output. Since then, we have switched from using a ppx_deriving rewriter to using the ppx_deriving compiler plugin, which has made the `show` plugin aware of the name of the current module. As a result, the `Ast`, `Ast_defs`, and `Aast_defs` module names are now printed in AAST pretty-printing output. This isn't necessary--these module prefixes are not needed to disambiguate any names (except for `Cnormal`, which can be disambiguated by context).

This change applies the `with_path = false` setting to the `show` configuration in `Ast`, `Ast_defs`, and `Aast_defs`. This causes them to produce the same output they did before we switched to the ppx_deriving compiler plugin, makes the TAST tests less noisy, and makes their output consistent with that of the `Aast` module (which uses the `with_path = false` setting).

Reviewed By: pittsw

Differential Revision: D7752638

fbshipit-source-id: bdd236921cce128336a4c62ce705d73753c0a7ac