Lexer tweak to disambiguate binop >= from type arg > followed by =
Summary:
In experimenting with new syntax, like `let` statements, we've come across `>=`
being ambiguously lexed; either it's the `>=` (greater-or-equal) token in a
binary expression, or it's the closing angle of a generic type argument,
followed by an assignment `=`, as in;

```
let x : vec<int> = vec[]; /* intended */
let y : vec<int>=vec[]; /* ambiguous */
```

We already had this problem for `>>` (either right-shift binop or two
consecutive closing angle brackets). Luckily, this only happens when types
"meet" expressions. It's easy to disambiguate.

Reviewed By: arun11111

Differential Revision: D7584327

fbshipit-source-id: 3c670cf0d44c7f2302a494f71d809276296d3e22