Use verify.py for full_fidelity_unit_tests
Summary:
The old system for full fidelity unit tests was quite confusing. This diff changes FFP to use verify system for tests.

I'm not 100% sure if what I'm testing is the same thing as what full_fidelity_unit_test and full_fidelity_ast_test were testing originally, but I think we're testing what matters:

Unfortunately, the FFP does not have a mode which gives errors when there are errors, and an s expression if there are none. So we'll have to test the s exp and the errors on both.
each file in full_fidelity_unit_test and full_fidelity_ast_test has a ".errors.exp" file and a ".sexp.exp" file, one which is the expected errors and one is the expected s expression. This looks like what the old ".exp" was doing, but with different results for different files.

With this change, the tests run much much faster, and we get a nice UI for editing the tests and seeing the difference.

Differential Revision: D7765961

fbshipit-source-id: b27a0c7a2fc58e7d83cd452f82cb55931274e442